Add feat: link to bulk discount index from merchant dashboard

+++ db/migrate/20220927120619_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
       t.integer :threshold
-      t.float :discount_percent
+      t.integer :discount_percent
       t.references :merchant, foreign_key: true
 
       t.timestamps
